`badRequest` outcome helper can take multiple expressions
In some situations, we want to report on problems that exist when two
resources conflict. In such a case, we want to emit an OperationOutcome
with an issue that has multiple `expression` attributes.

To that end, our `badRequest()` helper can now take an array of
expression strings as its second argument.

Implementation notes:
- `arrayify` was previously defined in `core.ts`, but `core.ts` imports
  from `outcomes.ts`, and so trying to pull from there could have
  created circular dependencies. Extracting the helper into a new
  `array.ts` lets both files depend on it safely. While there I also
  added test cases.
