+++ packages/core/src/outcomes.test.ts
@@ -89,6 +89,7 @@ describe('Outcomes', () => {
   test('Bad Request', () => {
     expect(isOk(badRequest('bad'))).toBe(false);
     expect(badRequest('bad', 'bad').issue?.[0]?.expression?.[0]).toBe('bad');
+    expect(badRequest('bad', ['bad[0]', 'bad[1]']).issue?.[0]?.expression).toEqual(['bad[0]', 'bad[1]']);
   });
 
   test.each([
