+++ packages/core/src/typeschema/crawler.test.ts
@@ -4,9 +4,10 @@ import { readJson } from '@medplum/definitions';
 import type { Attachment, Bundle, Coding, Observation, Patient, StructureDefinition } from '@medplum/fhirtypes';
 import { readFileSync } from 'node:fs';
 import { resolve } from 'node:path';
+import { arrayify } from '../array';
 import { LOINC } from '../constants';
 import { toTypedValue } from '../fhirpath/utils';
-import { arrayify, sleep } from '../utils';
+import { sleep } from '../utils';
 import { crawlTypedValue, crawlTypedValueAsync, pathToJSONPointer } from './crawler';
 import type { InternalTypeSchema } from './types';
 import { indexStructureDefinitionBundle, parseStructureDefinition } from './types';
